Move file, test more intenational, wording etc...
