+++ api/scripts/send-invitations-to-sco-organizations.js
@@ -6,7 +6,7 @@
 require('dotenv').config();
 const request = require('request-promise-native');
 
-const { findOrganizationsByExternalIds, organizeOrganizationsByExternalId } = require('./organizations-by-external-id-helper');
+const { findOrganizationsByExternalIds, organizeOrganizationsByExternalId } = require('./helpers/organizations-by-external-id-helper');
 const { parseCsv } = require('../scripts/helpers/csvHelpers');
 
 const baseUrl = process.env.BASE_URL || 'http://localhost:3000';
